fix: SQL components are pulled out to be shared components
