+++ ui/src/pages/DataAnalysis/components/SQLEditor/SqlTable/index.tsx
