+++ ui/src/pages/DataAnalysis/components/SQLEditor/index.less
@@ -1,4 +1,4 @@
-@import "../../../../../styles/var.less";
+@import "../../../../styles/var.less";
   .editorMain {
     display: flex;
     flex: 1;
